+++ src/lokalise/lokalise_api.ts
@@ -1,4 +1,5 @@
 import {
+	AuditLogs,
 	Branches,
 	Comments,
 	Contributors,
@@ -45,6 +46,13 @@ export class LokaliseApi extends BaseClient {
 		// Default to "api2" version if not explicitly provided
 		this.clientData.version = params.version ?? "api2";
 		this.clientData.authHeader = params.header ?? this.clientData.authHeader;
+	}
+
+	/**
+	 * Access Audit-related endpoints.
+	 */
+	auditLogs(): AuditLogs {
+		return new AuditLogs(this.clientData);
 	}
 
 	/**