+++ test/base_collection/base_collection.spec.ts
@@ -6,31 +6,31 @@ describe("BaseCollection", () => {
 	const collection = new DummyCollection(lokaliseApi.clientData);
 
 	it("should throw error if rootElementName is not defined", () => {
-		expect(() => collection.testRootElementName()).toThrowError(
+		expect(() => collection.testRootElementName()).toThrow(
 			"rootElementName is not defined. Subclasses must override `rootElementName`.",
 		);
 	});
 
 	it("should throw error if rootElementNameSingular is not defined", () => {
-		expect(() => collection.testRootElementNameSingular()).toThrowError(
+		expect(() => collection.testRootElementNameSingular()).toThrow(
 			"rootElementNameSingular is not defined. Subclasses must override `rootElementNameSingular`.",
 		);
 	});
 
 	it("should throw error if secondaryElementClass is not defined", () => {
-		expect(() => collection.testSecondaryElementClass()).toThrowError(
+		expect(() => collection.testSecondaryElementClass()).toThrow(
 			"Secondary elements are not supported by this collection",
 		);
 	});
 
 	it("should throw error if secondaryElementNameSingular is not defined", () => {
-		expect(() => collection.testSecondaryElementNameSingular()).toThrowError(
+		expect(() => collection.testSecondaryElementNameSingular()).toThrow(
 			"secondaryElementNameSingular is not defined. Subclasses must override this if secondary elements are used.",
 		);
 	});
 
 	it("should throw error if ", () => {
-		expect(() => collection.testGetUri(null)).toThrowError(
+		expect(() => collection.testGetUri(null)).toThrow(
 			"No URI or prefixURI provided.",
 		);
 	});