+++ test/keys/keys.spec.ts
@@ -159,6 +159,10 @@ describe("Keys", () => {
 				...params,
 			});
 		} catch (e) {
+			if (!(e instanceof Error)) {
+				throw e;
+			}
+
 			expect(e.message).toEqual(
 				"Expected an array under 'keys' for cursor pagination but received: object",
 			);
@@ -259,6 +263,10 @@ describe("Keys", () => {
 		try {
 			await lokaliseApi.keys().get(keyId, params);
 		} catch (e) {
+			if (!(e instanceof Error)) {
+				throw e;
+			}
+
 			expect(e.message).toEqual("Missing property 'key' in JSON object");
 		}
 	});